refactor: move more puddle stuff to Shared
Still can't do stuff that creates puddles :(
